requirements.txt
beautifulsoup4
faiss-cpu
google-generativeai
groq
langchain
langchain-community
langchain-google-genai
python-dotenv
requests
streamlit
